refactor: use new TableCard and Table components
