+++ src/components/ui/BrowsePage/Dashboard/RecentDataLinksCard.tsx
@@ -1,9 +1,8 @@
 import { Typography } from '@material-tailwind/react';
 
 import DashboardCard from '@/components/ui/BrowsePage/Dashboard/FgDashboardCard';
-import ProxiedPathRow from '@/components/ui/LinksPage/ProxiedPathRow';
-import { TableRow } from '@/components/ui/widgets/TableCard';
-import { TableRowSkeleton } from '@/components/ui/widgets/Loaders';
+import { linksColumns } from '@/components/ui/Table/linksColumns';
+import { Table } from '@/components/ui/Table/TableCard';
 import { useProxiedPathContext } from '@/contexts/ProxiedPathContext';
 
 export default function RecentDataLinksCard() {
@@ -14,16 +13,7 @@ export default function RecentDataLinksCard() {
 
   return (
     <DashboardCard title="Recently created data links">
-      {loadingProxiedPaths ? (
-        Array(5)
-          .fill(0)
-          .map((_, index) => (
-            <TableRowSkeleton
-              key={index}
-              gridColsClass="grid-cols-[1.5fr_2.5fr_1.5fr_1fr]"
-            />
-          ))
-      ) : recentDataLinks.length === 0 ? (
+      {recentDataLinks.length === 0 ? (
         <div className="px-4 pt-4 flex flex-col gap-4">
           <Typography className="text-muted-foreground">
             No data links created yet.
@@ -38,14 +28,13 @@ export default function RecentDataLinksCard() {
           </Typography>
         </div>
       ) : (
-        recentDataLinks.map(proxiedPath => (
-          <TableRow
-            gridColsClass="grid-cols-[1.5fr_2.5fr_1.5fr_1fr]"
-            key={proxiedPath.sharing_key}
-          >
-            <ProxiedPathRow key={proxiedPath.sharing_key} item={proxiedPath} />
-          </TableRow>
-        ))
+        <Table
+          columns={linksColumns}
+          data={recentDataLinks || []}
+          gridColsClass="grid-cols-[1.5fr_2.5fr_1.5fr_1fr_1fr]"
+          loadingState={loadingProxiedPaths}
+          emptyText="No shared paths."
+        />
       )}
     </DashboardCard>
   );
